# Swagger/Swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OneCard + Product API",
    "version": "1.0.0",
    "description": "Corrected API documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8597",
      "description": "Local Server"
    }
  ],
  "paths": {
...
    "/api/update/{id}": {
      "put": {
        "summary": "Update user profile",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUser" }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateUser": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
...
    }
  }
}